Move Web.HttpApiData content to .Internal module and export only classes
